fix(devops): resolve blue green improvements
