+++ docker-compose.web.prod.yml
@@ -1,25 +1,48 @@
+x-web-service: &web-service
+  build:
+    context: .
+    dockerfile: apps/web/Dockerfile
+    target: runner
+    secrets:
+      - web_env
+  env_file:
+    - path: apps/web/.env.local
+      required: true
+  environment: &web-environment
+    HOSTNAME: "0.0.0.0"
+    NEXT_TELEMETRY_DISABLED: "1"
+    NODE_ENV: production
+    PORT: "7803"
+    SUPABASE_SERVER_URL: ${DOCKER_INTERNAL_SUPABASE_URL:-http://host.docker.internal:8001}
+  extra_hosts:
+    - "host.docker.internal:host-gateway"
+  init: true
+
 services:
   web:
-    build:
-      context: .
-      dockerfile: apps/web/Dockerfile
-      target: runner
-      secrets:
-        - web_env
-    env_file:
-      - path: apps/web/.env.local
-        required: true
-    environment:
-      HOSTNAME: "0.0.0.0"
-      NEXT_TELEMETRY_DISABLED: "1"
-      NODE_ENV: production
-      PORT: "7803"
-      SUPABASE_SERVER_URL: ${DOCKER_INTERNAL_SUPABASE_URL:-http://host.docker.internal:8001}
-    extra_hosts:
-      - "host.docker.internal:host-gateway"
+    <<: *web-service
+    ports:
+      - "7803:7803"
+
+  web-blue:
+    <<: *web-service
+
+  web-green:
+    <<: *web-service
+
+  web-proxy:
+    healthcheck:
+      test: ["CMD-SHELL", "wget -q -O /dev/null http://127.0.0.1:7803/api/health || exit 1"]
+      interval: 30s
+      timeout: 5s
+      start_period: 10s
+      retries: 3
+    image: nginx:1.27-alpine
     init: true
     ports:
       - "7803:7803"
+    volumes:
+      - ./tmp/docker-web/prod/nginx.conf:/etc/nginx/conf.d/default.conf:ro
 
   redis:
     profiles: ["redis"]
@@ -41,7 +64,7 @@ services:
       SRH_CONNECTION_STRING: redis://redis:6379
       SRH_MODE: env
       # Export SRH_TOKEN from your shell or CI before enabling the production Redis profile.
-      SRH_TOKEN: ${SRH_TOKEN}
+      SRH_TOKEN: ${SRH_TOKEN:?SRH_TOKEN must be set when enabling the redis profile}
     image: hiett/serverless-redis-http:latest
     ports:
       - "8079:80"
