+++ package.json
@@ -45,7 +45,9 @@
     "dev:web:docker": "node scripts/docker-web.js up",
     "dev:web:docker:down": "node scripts/docker-web.js down",
     "serve:web:docker": "node scripts/docker-web.js up --mode prod",
+    "serve:web:docker:bg": "node scripts/docker-web.js up --mode prod --strategy blue-green",
     "serve:web:docker:down": "node scripts/docker-web.js down --mode prod",
+    "serve:web:docker:bg:down": "node scripts/docker-web.js down --mode prod --strategy blue-green",
     "dev:webns": "turbo run dev -F @tuturuuu/types -F @tuturuuu/supabase",
     "dev:external": "turbo run dev -F @tuturuuu/external",
     "dev:calendar": "turbo run dev -F @tuturuuu/types -F @tuturuuu/supabase -F @tuturuuu/calendar -F @tuturuuu/web",
